Refactor: Implement data layer separation and refactoring
This commit implements the data layer separation, Edge function refactoring, and component refactoring as outlined in the refactoring plan. It includes the creation of a template configuration module, refactoring of the prompt builder, and component-specific refactoring.
