+++ supabase/functions/generate-resume/index.ts
@@ -1,4 +1,3 @@
-
 import "https://deno.land/x/xhr@0.1.0/mod.ts";
 import { serve } from "https://deno.land/std@0.168.0/http/server.ts";
 
@@ -9,7 +8,7 @@ const corsHeaders = {
   'Access-Control-Allow-Headers': 'authorization, x-client-info, apikey, content-type',
 };
 
-// Template tone profiles for dynamic prompt generation
+// Centralized template tone profiles
 const toneProfiles = {
   modern: {
     writingStyle: 'professional',
